Refactor VisionTarget sorting
Previously, there was a dead zone where two targets could be considered
equal if the area difference between them was less than 1 because the
double difference would be cast to an int (0), despite the two things
not ACTUALLY being equal.

It’s a very small thing, and, honestly, targets that close in size are
probably effectively equal, but we might as well return things based on
data and with as little “dead zone” as possible.